feat(step3): radio recovery + hardware watchdog
SX1262 fault detection and guarded recovery:
  radio_guard_reset(reason): backoff wrapper (min RADIO_RESET_BACKOFF_MS=10s
    between consecutive hard resets); increments radio_reset_backoff when
    denied; logs reason string every time
  radio_check_timeouts(): per-loop-iteration check; fires radio_rx_timeout
    metric + guarded reset when RX silent >= RADIO_RX_SILENCE_MS (60s)
  BUSY-stuck streak: s_busy_stuck_streak; after RADIO_BUSY_STUCK_MAX (3)
    consecutive stalls → radio_guard_reset(busy_stuck); streak cleared on
    any successful BUSY-low; per-event WARN log-rate-limited (1,2,4,8,…)
  TX HW-timeout / deadline: radio_transmit() already tracked s_tx_fail_streak;
    now calls radio_guard_reset(tx_timeout) instead of radio_hard_reset(),
    honouring the backoff cooldown
  Spurious-IRQ streak: radio_service_rx() → radio_guard_reset(spurious_irq)

New metrics fields (rivr_metrics_t):
  radio_rx_timeout    — RX silent >60s → soft reset fired
  radio_reset_backoff — hard reset denied by 10s backoff cooldown

@MET JSON expanded: added rx_tout, rst_bkof keys (18 total now)

Hardware watchdog (ESP builds only):
  esp_task_wdt_add(NULL) at boot (main task subscribes)
  esp_task_wdt_reset() called once per main-loop iteration
  timeout=30s configured in sdkconfig.defaults
  IDLE tasks excluded (may be starved during long LoRa TX windows)
  Entire WDT block is #ifndef RIVR_SIM_MODE guarded

radio_sx1276.c: added radio_check_timeouts() no-op stub so the shared
  main.c call-site links for the LilyGo SX1276 build

Fault injection (RIVR_FAULT_INJECT=1):
  g_fault_busy_stuck  — makes platform_sx1262_wait_busy return false
  g_fault_tx_no_done  — suppresses TxDone IRQ flag in TX poll loop
  g_fault_rx_silence  — suppresses DIO1 event dispatch
  radio_fault_reset_state() — resets all statics + fault flags + metrics

Test suite:
  tests/Makefile — Linux/macOS Makefile; 'make -C tests' runs both suites
  tests/test_radio_recovery.c — 5 test cases exercising every recovery path:
    TC-1  BUSY stuck × 3 → stall counter + hard reset (24 assertions)
    TC-2  Backoff denies rapid re-reset
    TC-3  Backoff expires after 10s cooldown; second reset fires
    TC-4  RX silence ≥ 60s → radio_rx_timeout + guarded reset
    TC-5  TX HW-timeout × 3 streak → guarded reset (tx_timeout)
  tests/include/ — host-build stubs: esp_attr.h, esp_log.h, esp_task_wdt.h,
    esp_idf_version.h, driver/gpio.h, driver/spi_master.h,
    freertos/FreeRTOS.h, freertos/task.h

Build:
  pio run -e repeater_esp32devkit_e22_900 → SUCCESS  (~20s)
  pio run -e client_esp32devkit_e22_900   → SUCCESS  (~20s)
  pio run -e repeater_lilygo_lora32_v21  → SUCCESS  (~13s)
  make -C tests → acceptance 45/45 PASS + recovery 24/24 PASS

+++ tests/test_stubs.c
@@ -32,14 +32,15 @@ void timebase_tick_hook(void)
 
 uint32_t tb_lmp_advance(uint32_t sender_tick)
 {
-    uint32_t local;
+    uint_fast32_t local;
     do {
         local = atomic_load_explicit(&g_lmp_tick, memory_order_relaxed);
-        uint32_t desired = (sender_tick > local ? sender_tick : local) + 1u;
+        uint_fast32_t desired = (sender_tick > local
+                                 ? (uint_fast32_t)sender_tick : local) + 1u;
         if (atomic_compare_exchange_weak_explicit(
                 &g_lmp_tick, &local, desired,
                 memory_order_release, memory_order_relaxed))
-            return desired;
+            return (uint32_t)desired;
     } while (1);
 }
